sprint-3/step-1 feat: вынести форму авторизации в компонент
